# api/docs/api/openapi.yaml
openapi: 3.0.0
components:
...
  schemas:
...
    TournamentType:
      enum:
        - SINGLE_GROUP
      type: string
    TeamApiModel:
      properties:
        short_name:
          type: string
        full_name:
          type: string
        description:
          type: string
        country:
          type: string
        logo_image_file:
          type: string
        id:
          type: number
          format: double
      required:
        - short_name
        - full_name
        - description
        - country
      type: object
      additionalProperties: false
    TournamentApiModel:
      properties:
        name:
          type: string
        description:
          type: string
        country:
          type: string
        type:
          $ref: "#/components/schemas/TournamentType"
        start_date:
          type: string
        end_date:
          type: string
        started:
          type: boolean
        ended:
          type: boolean
        winner_id:
          type: number
          format: double
        teams:
          items:
            $ref: "#/components/schemas/TeamApiModel"
          type: array
        id:
          type: number
          format: double
      required:
        - name
        - description
        - country
        - type
        - start_date
        - end_date
        - started
        - ended
      type: object
      additionalProperties: false
    ItemsWithPagination_TournamentApiModel_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/TournamentApiModel"
          type: array
        total:
          type: number
          format: double
      required:
        - items
        - total
      type: object
      additionalProperties: false
    MatchApiModel:
      properties:
        date:
          type: string
        tournament_id:
          type: number
          format: double
        team1_id:
          type: number
          format: double
        team2_id:
          type: number
          format: double
        type:
          type: string
        team1_score:
          type: number
          format: double
        team2_score:
          type: number
          format: double
        included_on_standings:
          type: boolean
        started:
          type: boolean
        finished:
          type: boolean
        winner_id:
          type: number
          format: double
        id:
          type: number
          format: double
      required:
        - date
        - tournament_id
        - team1_id
        - team2_id
        - type
        - team1_score
        - team2_score
        - included_on_standings
        - started
        - finished
      type: object
      additionalProperties: false
    ItemsWithPagination_MatchApiModel_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/MatchApiModel"
          type: array
        total:
          type: number
          format: double
      required:
        - items
        - total
      type: object
      additionalProperties: false
    StandingsApiModel:
      properties:
        wins:
          type: number
          format: double
        losses:
          type: number
          format: double
        maps_won:
          type: number
          format: double
        maps_lost:
          type: number
          format: double
        rounds_won:
          type: number
          format: double
        rounds_lost:
          type: number
          format: double
        tournament_id:
          type: number
          format: double
        team_id:
          type: number
          format: double
        position:
          type: number
          format: double
        id:
          type: number
          format: double
      required:
        - wins
        - losses
        - maps_won
        - maps_lost
        - rounds_won
        - rounds_lost
        - tournament_id
        - team_id
        - position
      type: object
      additionalProperties: false
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
info:
  title: backend
  version: 0.1.0
  contact: {}
paths:
...
  /tournaments:
    get:
      operationId: getTournaments
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsWithPagination_TournamentApiModel_"
      description: Retrieves all tournaments with optional pagination
      security: []
      parameters:
        - description: Maximum number of tournaments to return
          in: query
          name: limit
          required: false
          schema:
            default: 10
            format: double
            type: number
        - description: Number of tournaments to skip
          in: query
          name: offset
          required: false
          schema:
            default: 0
            format: double
            type: number
    post:
      operationId: createTournament
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentApiModel"
      description: Creates a new tournament
      security: []
      parameters: []
      requestBody:
        description: The tournament data to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentApiModel"
              description: The tournament data to create
  /tournaments/{tournamentId}:
    get:
      operationId: getTournament
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentApiModel"
      description: Retrieves a specific tournament by its ID
      security: []
      parameters:
        - description: The ID of the tournament to retrieve
          in: path
          name: tournamentId
          required: true
          schema:
            format: double
            type: number
    put:
      operationId: updateTournament
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentApiModel"
      description: Updates an existing tournament
      security: []
      parameters:
        - description: The ID of the tournament to update
          in: path
          name: tournamentId
          required: true
          schema:
            format: double
            type: number
      requestBody:
        description: The tournament data to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentApiModel"
              description: The tournament data to update
  /tournaments/{tournamentId}/schedule:
    get:
      operationId: getTournamentSchedule
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsWithPagination_MatchApiModel_"
      description: Retrieves the match schedule for a tournament
      security: []
      parameters:
        - description: The ID of the tournament to retrieve the schedule for
          in: path
          name: tournamentId
          required: true
          schema:
            format: double
            type: number
        - description: Maximum number of matches to return
          in: query
          name: limit
          required: false
          schema:
            format: double
            type: number
        - description: Number of matches to skip
          in: query
          name: offset
          required: false
          schema:
            format: double
            type: number
  /tournaments/{tournamentId}/standings:
    get:
      operationId: getTournamentStandings
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/StandingsApiModel"
                type: array
      description: Retrieves the standings for a tournament
      security: []
      parameters:
        - description: The ID of the tournament to retrieve standings for
          in: path
          name: tournamentId
          required: true
          schema:
            format: double
            type: number
  /tournaments/{tournamentId}/start:
    post:
      operationId: startTournament
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentApiModel"
      description: Starts a tournament
      security: []
      parameters:
        - description: The ID of the tournament to start
          in: path
          name: tournamentId
          required: true
          schema:
            format: double
            type: number
  /tournaments/{tournamentId}/end:
    post:
      operationId: endTournament
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentApiModel"
      description: Ends a tournament
      security: []
      parameters:
        - description: The ID of the tournament to end
          in: path
          name: tournamentId
          required: true
          schema:
            format: double
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                winner_id:
                  type: number
                  format: double
              required:
                - winner_id
              type: object
...
servers:
  - url: /api